flask_app/requirements.txt
Flask
requests
python-firebase